ifcparse: widen all integer attribute types to int64_t for consistency
Follow-up to the scalar-only fix in #8754, per aothms's direct request on
that PR ("Please do make all int types consistent") and his own original
2023 design intent on issue #3058 ("make all integers (incl. schema
namespaces) an int64_t"). Widens the remaining inconsistent spots now that
compatibility isn't a constraint on this v0.9 branch:

- Integer aggregates (IfcTriangulatedFaceSet.CoordIndex and similar
  List<int> attributes), including the SWIG to_vec_int/to_vec_vec_int
  helpers, which previously silently truncated via static_cast<int> on the
  Python-set path - the same bug class as the original scalar issue.
- The schema code generator (express/mapping.py's integer type mapping),
  and all 12 generated schema header/source pairs regenerated to match, so
  every schema-typed getter/setter (e.g. IfcOwnerHistory::CreationDate) is
  int64_t end to end, not just the dynamic attribute-value path.

Instance/reference identifiers (STEP #123 ids) are deliberately left at
32-bit: they're a file-local index into internal maps, not an EXPRESS
domain value an application chooses, and no realistic STEP file has
billions of entities. The lexer's Token_IDENTIFIER parsing still funnels
through a 32-bit int for this reason - flagged as a known, low-risk gap
rather than fixed, since fixing it would mean touching indexing/hashing
code for no realistic benefit.

Verified: original PR's round-trip tests extended with aggregate cases
(IfcTriangulatedFaceSet.CoordIndex, InnerCoordIndices) at 64-bit boundary
values, in memory and through STEP text, IFC2X3 and IFC4. A standalone C++
program exercising the generated schema API directly (Ifc4::IfcOwnerHistory
::setCreationDate/CreationDate, IfcTriangulatedFaceSet::setCoordIndex/
CoordIndex) confirms int64_t end to end, bypassing SWIG. Full build
(BUILD_IFCGEOM, WITH_OPENCASCADE, BUILD_IFCPYTHON, IFC2X3+IFC4) clean.
test/util/test_attribute.py and test_file.py pass unchanged.

This contribution was produced with the assistance of an AI coding tool.

+++ src/helpers/pset.cpp
@@ -88,7 +88,7 @@ property_value from_attribute(const attribute_value& value) {
 
     switch (value.type()) {
     case ifcopenshell::Argument_INT:
-        return static_cast<int>(value);
+        return static_cast<int64_t>(value);
     case ifcopenshell::Argument_BOOL:
         return static_cast<bool>(value);
     case ifcopenshell::Argument_LOGICAL: {
@@ -120,7 +120,7 @@ property_value from_attribute(const attribute_value& value) {
         return instance;
     }
     case ifcopenshell::Argument_AGGREGATE_OF_INT:
-        return scalar_list(static_cast<std::vector<int>>(value));
+        return scalar_list(static_cast<std::vector<int64_t>>(value));
     case ifcopenshell::Argument_AGGREGATE_OF_DOUBLE:
         return scalar_list(static_cast<std::vector<double>>(value));
     case ifcopenshell::Argument_AGGREGATE_OF_STRING:
